Background worker
Switch to the full source of Pirple. Now server is not in index.js. It's in ./lib/server.js. Now app uses config.env (default is staging when started without special parameter). Minor changes in handler.js. I.e. query is now queryStringObject in a get or delete method. The app inform the user with a sms, if one of his checks changes and perform the life checks of the specific url every minute.